chapter_003/src/test/java/ru/job4j/generics/ConvertListTest.java
package ru.job4j.generics;

import org.junit.Test;

import java.util.Arrays;
import java.util.LinkedList;
import java.util.List;

import static org.hamcrest.core.Is.is;
import static org.junit.Assert.assertThat;

/**
 * Test.
 *
 * @author Yury Doronin (doronin.ltd@gmail.com)
 * @version 1.0
 * @since 24.04.2019
 */
public class ConvertListTest {

    /**
     * Test convert.
     */
    @Test
    public void whenListIntThenListInteger() {
        ConvertList converter = new ConvertList();
        List<int[]> list = new LinkedList<>();
        list.add(new int[]{1, 2});
        list.add(new int[]{3, 4, 5, 6});
        List<Integer> result = converter.convert(list);
        List<Integer> expect = Arrays.asList(1, 2, 3, 4, 5, 6);
        assertThat(result, is(expect));
    }
}